Fix default value of LangID.

+++ src/Documenter/Msdn/MsdnDocumenterConfig.cs
@@ -418,7 +418,7 @@ namespace NDoc.Documenter.Msdn
 		/// <remarks>The language ID of the locale used by the compiled helpfile</remarks>
 		[Category("HTML Help Options")]
 		[Description("The ID of the language the help file is in.")]
-		[DefaultValue(1033)]
+		[DefaultValue((short)1033)]
 		[Editor(typeof(LangIdEditor), typeof(UITypeEditor))]
 		public short LangID
 		{
